requirements.txt
requests
beautifulsoup4

neo4j

python-dotenv

ipykernel
jupyter
pandas

rapidfuzz